requirements.txt
Django>=4.2
djangorestframework
google-cloud-pubsub
